+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Laravel 8 support
 ### Changed
 - Change `emmit` method name to `dispatch`
+- Change `assertEmitted` method name to `assertDispatched`
 ## [v1.6.0]
 ### Added
 - Added RPCs test