+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Properties on publish method
 ### Added
 - Laravel 8 support
+- Added `Pigeon::assertNotDispatched()` method to Fake driver
 ### Changed
 - Change `emmit` method name to `dispatch`
 ## [v1.6.0]